clean train script
