+++ onmt/translate/translator.py
@@ -18,11 +18,6 @@ from onmt.utils.misc import tile, set_random_seed, report_matrix
 from onmt.utils.alignment import extract_alignment, build_align_pharaoh
 from onmt.modules.copy_generator import collapse_copy_scores
 
-from onmt.dynamicdata.config import read_data_config, verify_shard_config
-from onmt.dynamicdata.transforms import set_train_opts
-from onmt.dynamicdata.vocab import load_fields, load_transforms
-from onmt.dynamicdata.iterators import TransformReader
-
 
 def build_translator(opt, report_score=True, logger=None, out_file=None):
     if out_file is None:
@@ -236,20 +231,8 @@ class Translator(object):
             report_score (bool) : See :func:`__init__()`.
             logger (logging.Logger or NoneType): See :func:`__init__()`.
         """
-
-        if opt.transforms_from_task is None:
-            assert opt.data_config is None
-            src_reader = inputters.str2reader[opt.data_type].from_opt(opt)
-            transforms_from_task = None
-        else:
-            assert opt.data_config is not None
-            transforms_from_task = opt.transforms_from_task
-            data_config = read_data_config(opt.data_config)
-            verify_shard_config(data_config)
-            transform_models, transforms = load_transforms(data_config)
-            set_train_opts(data_config, transforms)
-            #fields = load_fields(data_config)  # using fields from model
-            src_reader = TransformReader(transforms_from_task, transforms[transforms_from_task])
+        # TODO: maybe add dynamic part
+        src_reader = inputters.str2reader[opt.data_type].from_opt(opt)
         tgt_reader = inputters.str2reader["text"].from_opt(opt)
         return cls(
             model,
